pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cshel-image-tools"
version = "0.1.0"
description = "MCP server for Google's Nano Banana Pro (Gemini 3 Pro Image): generate, edit, compose, and upscale images."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Chris Shelton" },
]
keywords = ["mcp", "model-context-protocol", "gemini", "nano-banana", "image-generation", "google-genai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp>=1.2.0",
    "google-genai>=1.0.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
]

[project.urls]
Homepage = "https://github.com/seashell86/cshel-image-tools"
Issues = "https://github.com/seashell86/cshel-image-tools/issues"

[project.scripts]
cshel-image-tools = "cshel_image_tools.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/cshel_image_tools"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "N"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = ["google.genai.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
